Move SDF utilities to importable package
