+++ INCHI-1-TEST/consumers.py
@@ -2,18 +2,8 @@ import ctypes
 import random
 from typing import Callable
 from pathlib import Path
-from sdf_pipeline import drivers
+from sdf_pipeline import drivers, utils
 from .inchi_api import make_inchi_from_molfile_text, get_inchi_key_from_inchi
-
-try:
-    # Optional import of RDKit.
-    from rdkit import Chem
-    from rdkit import RDLogger
-
-    # Suppress RDKit console output.
-    RDLogger.DisableLog("rdApp.*")
-except ImportError:
-    pass
 
 
 def regression_consumer(
@@ -43,33 +33,6 @@ def regression_consumer(
     )
 
 
-def _permute(indices: list[int]) -> list[int]:
-    # Shuffle without in-place mutation.
-    # See https://docs.python.org/3/library/random.html#random.shuffle.
-    return random.sample(indices, len(indices))
-
-
-def permute_molblock(molblock: str) -> str | None:
-    mol = Chem.MolFromMolBlock(
-        molblock, sanitize=False, removeHs=False, strictParsing=False
-    )
-    if mol is None:
-        return None
-
-    atom_indices = [atom.GetIdx() for atom in mol.GetAtoms()]
-    if not atom_indices:
-        return None
-
-    atom_indices_permuted = _permute(atom_indices)
-    if len(atom_indices) > 1:
-        # Enforce different permutation of atom indices.
-        while atom_indices_permuted == atom_indices:
-            atom_indices_permuted = _permute(atom_indices)
-    mol_permuted = Chem.RenumberAtoms(mol, atom_indices_permuted)
-
-    return Chem.MolToMolBlock(mol_permuted, kekulize=False)
-
-
 def invariance_consumer(
     molfile: str,
     get_molfile_id: Callable,
@@ -85,7 +48,7 @@ def invariance_consumer(
     random.seed(42)
 
     for _ in range(n_invariance_runs):
-        molfile_permuted = permute_molblock(molfile)
+        molfile_permuted = utils.permute_molblock(molfile)
         if molfile_permuted is None:
             # Failure to parse `molfile`,
             # return empty result which will be handled upstream.
