+++ INCHI-1-TEST/src/sdf_pipeline/core.py
@@ -1,37 +1,18 @@
 import multiprocessing
 from queue import Empty
-import gzip
 from typing import Callable, TYPE_CHECKING
 from collections.abc import Generator
-from sdf_pipeline import logger
+from sdf_pipeline import logger, utils
 
 if TYPE_CHECKING:
     # https://adamj.eu/tech/2021/05/13/python-type-hints-how-to-fix-circular-imports/
     from sdf_pipeline.drivers import ConsumerResult
 
 
-def read_records_from_gzipped_sdf(sdf_path: str) -> Generator[str, None, None]:
-    # https://en.wikipedia.org/wiki/Chemical_table_file#SDF"
-    current_record = ""
-    # TODO: guard file opening.
-    with gzip.open(sdf_path, "rb") as gzipped_sdf:
-        # Decompress SDF line-by-line to avoid loading entire SDF into memory.
-        for decompressed_line in gzipped_sdf:
-            decoded_line = decompressed_line.decode("utf-8", "backslashreplace")
-            current_record += decoded_line
-            if decoded_line.strip() == "$$$$":
-                # TODO: harden SDF parsing according to
-                # http://www.dalkescientific.com/writings/diary/archive/2020/09/18/handling_the_sdf_record_delimiter.html.
-                yield current_record
-                current_record = ""
-
-    return None
-
-
 def _produce_molfiles(
     molfile_queue: multiprocessing.Queue, sdf_path: str, n_poison_pills: int
 ) -> None:
-    for molfile in read_records_from_gzipped_sdf(sdf_path):
+    for molfile in utils.read_records_from_gzipped_sdf(sdf_path):
         molfile_queue.put(molfile)
 
     for _ in range(n_poison_pills):
